Implement same authorisation logic for OCPP 2.0.1
Same as that used for OCPP 1.6, based on configuration.yaml
